Look at the TLDR example using memoization - is this valid?
It seemed odd to me that this value is not memoized but instead
the object is recreated on every access. Is this intentional
because memoization (for some reason) does not work here or does
memoization work as well?

(Decided to PR as this is easier to discuss and maybe easier to
 fix then)