Support for Unconventional Transition Associations
This commit adds support for an `association_name` option when
initializing a state-machine. This supports transition model
associations where the transition model's table-name doesn't match the
association--for example, when models are namespaced.

Additionally, the model_foreign_key, transition1_join, and
transition2_join methods rely on the transition class's table-name
matching the name used to associate it to the parent model. Instead, a
reflection is performed on the transition association, and is used to
retrieve the correct table and foreign-key values.

See gocardless/statesman#130 for motivation and other details.