Handle custom transition association names gracefully
Previously, statesman required the association with the transition class to use
the exact name of the transition class, unless a `transition_name` was specified
along with the `transition_class`.

This PR updates the way we look for the association with the transition class so
as `transition_name` no longer needs to be overridden to work with custom named
associations.