allowed_transitions? accept metadata to pass to guards
Internally, allowed_transitions? uses can_transition_to? to call all guards on a transition to check it's feasability.

While you can pass metadata into can_transition_to? for the guards to use, you don't have the possibility on allowed_transitions?

This allows you to list the allowed transitions given a metadata context (we use metadata in our guards heavily to add a bit of context of the change)

There are no breaking changes.