Enable inheritance between state machine classes
In order to use a state machine for tracking the various states of your application object, Statesman requires that you implement state machine classes. These include the Statesman::Machine module and make use of it's class methods to define a  set of valid states, transitions and callbacks.

Depending on the application domain, it might be required that you implement different state machine classes for the same object based on it's type. As these classes are bound to share some states, transitions and/or callbacks, this inherently results in duplicating the code which defines these states, transitions and/or callbacks in each of the state machine classes.

Since the calls to define the states, transitions and callbacks of a state machine class are executed when Ruby loads the file containing the class, it is not possible to use Ruby's inbuilt inheritance mechanism to share these definitions between classes.

Here, we introduce a Module that facilitates the sharing of such definitions between state machine classes. This can be included within any state machine class and it's functions can be invoked at load time to inherit the definitions of states, transitions and callbacks from another state machine class.