Make after_commit work reliabliy inside nested transactions
In the current implementation, after_commit callbacks are ran
immediately after the `ActiveRecord::Base.transaction` block,
which is not guaranteed to be after the transaction has been
committed to the database.

Speficially, if statesman code is running inside another transaction,
the after_commit block will be running prematurely.

This commit fixes that behavious by using ActiveRecord’s
add_transaction_record method (which is how after_commit
callbacks are implemented in Rails).