Update bundler to ~> 2.0.1
We previously pinned bundler 1.3 which is quite old right now. We likely
avoided using 2.0.1 because it requires Ruby >= 2.3, but given Ruby 2.2
is now EoL we can drop this requirement.

This commit will be followed by dropping of EoL Ruby versions from
official support. We also add a .ruby-version file that tracks the
latest supported Ruby version for ease of development.