Handle transitions differently for MySQL (#399)
* Inverse the order of writes on statesman transitions

StatesMan will update the old transitions to `most_recent:
nil` and then insert the new transition as `most_recent: true`.
However that can cause deadlocks on MySQL when running concurrent state
transitions.
On the update, MySQL will hold a gap lock to the unique indexes there are, so
other transactions cannot insert. After two updates on the most_recent
and two inserts, the second insert will fail with a deadlock.
For more explanation see the PR description that includes this commit.

* Make sure the initial most_recent state is false/nil

* Lock-lesser transitions without additional queries

https://github.com/gocardless/statesman/pull/350

@arthurnn opened #350 to reduce the impact of Statesman taking gap locks
when using MySQL. The crux of the issue is that MySQL's implementation
of REPEATABLE READ can take wide locks when an update touches no rows,
which happens frequently on the first transition of Statesman.

By first creating the new transition, we can avoid issuing an update
that will take the large gap lock. This order of queries meant we added
an additional query to the transition step which could impact people who
rely on low-latency Statesman transitions.

This commit is another take on the same approach that collapses two
queries into one, by taking the update of the old and new transition's
most_recent column and updating them together. It's slightly janky but
if robust, would be a good alternative to avoid additional latency.

* fmt

* Replace Arel#or with raw SQL

* Attempt to fix typecasting

* Fix MySQL issue with ordered updates

This commit first refactors the construction of transition SQL
statements to use Arel. This is safer and hopefully more readable than
constructing large SQL statements using strings.

It also fixes a bug with transition updates where MySQL would throw
index violations. This was caused by MySQL validating index constraints
across a partially applied update, where Statesman would set the latest
transition's most_recent = 't' before unsetting the previous.

* Smooth-over breaking API change in Arel

https://github.com/rails/rails/commit/7508284800f67b4611c767bff9eae7045674b66f

When merged with Rails core, Arel removed an initialisation parameter
that was unused in all but a few of the Arel features. For now, provide
a helper local to the ActiveRecord adapter that can construct Manager
classes independent of the API change.

* Handle transitions differently for MySQL
To avoid a gap lock with MySQL, first insert the new transition with most_recent=false and then flip the values for the latest two transitions in order.

* MySQL specific gaplock protection is configurable
Instead of depending on the name of the ActiveRecord adapter to identify if the database is MySQL, gaplock protection is configurable, as suggested by @jurre on #399

* Loosen dependency on pg gem
This reflects a recent change in Rails (https://github.com/rails/rails/commit/592358e182effecebe8c6a4645bd4431f5a73654)

Co-authored-by: Arthur Neves <arthurnn@gmail.com>
Co-authored-by: Lawrence Jones <lawrjone@gmail.com>
Co-authored-by: Grey Baker <greysteil@gmail.com>