Add from state method to transition history (#493)
This is a continuation of the work of
[marclerodrigues](https://github.com/marclerodrigues) in [No easy way to
find previous state](https://github.com/gocardless/statesman/pull/348).

One of the concerns was:
> I like the idea.
>
> Is it possible to test if this gracefully falls back when a table does
not have the from_state column?
> It'd be great if we didn't make table migrations required.

This PR adds checks to see if the column exists before querying it,
therefore users who have not added this column will be able to continue
with existing behaviour. This means that this is not a breaking change.

Persisting this information means that accessing it is faster and easier
and avoid problems such as mentioned in [this
comment](https://github.com/gocardless/statesman/issues/263#issuecomment-858770174)
and the issue containing it.

---------

Co-authored-by: Joe Southan <JoeSouthan@users.noreply.github.com>